refactor: FCM 연동 클래스명 변경 (NotificationEventListener -> NotificationEventClient) (#26)
- 클래스의 실제 역할에 알맞도록 `NotificationEventListener`를 `NotificationEventClient`로 이름 변경
